feat(primitives): move AnnouncePolicy from configuration to primitives (#1795)
Fixes the inverted dependency where torrust-tracker-primitives depended on
torrust-tracker-configuration for AnnouncePolicy. After this change:

- Before: primitives -> configuration (inverted dep)
- After:  configuration -> primitives (natural direction)

Changes:
- Move AnnouncePolicy struct+impls into packages/primitives/src/announce.rs
- Export AnnouncePolicy from torrust_tracker_primitives lib.rs
- Remove torrust-tracker-configuration dep from primitives/Cargo.toml
- Add torrust-tracker-primitives dep to configuration/Cargo.toml
- Replace AnnouncePolicy definition in configuration/src/lib.rs with a
  #[deprecated] re-export for external consumer backwards compatibility
- Update configuration/src/v2_0_0/core.rs to import from torrust_tracker_primitives
- Update bittorrent-http-tracker-protocol test import + remove unused
  torrust-tracker-configuration dep from its Cargo.toml
- Update bittorrent-tracker-core integration test import
- Update announce_handler.rs doc comment import path

All acceptance criteria verified:
- cargo build --workspace: clean
- cargo test --workspace: all pass
- linter all: exit code 0
- cargo machete: no unused deps
- grep torrust_tracker_configuration::AnnouncePolicy: zero matches

Closes #1795
Part of #1669 (SI-04)
