+++ packages/http-protocol/Cargo.toml
@@ -25,7 +25,6 @@ serde = { version = "1", features = [ "derive" ] }
 serde_bencode = "0"
 thiserror = "2"
 torrust-tracker-clock = { version = "3.0.0-develop", path = "../clock" }
-torrust-tracker-configuration = { version = "3.0.0-develop", path = "../configuration" }
 torrust-tracker-contrib-bencode = { version = "3.0.0-develop", path = "../../contrib/bencode" }
 torrust-tracker-located-error = { version = "3.0.0-develop", path = "../located-error" }
 torrust-tracker-primitives = { version = "3.0.0-develop", path = "../primitives" }
